Refactor Modal JS into standalone file, remove duplicated JS and HTML
