+++ resources/views/hardware/checkout.blade.php
@@ -55,7 +55,7 @@
               </div>
               <div class="col-md-1 col-sm-1 text-left">
                   @can('create', \App\Models\User::class)
-                    <a href='#' data-toggle="modal"  data-target="#createModal" data-dependency="user" data-select='assigned_user' class="btn btn-sm btn-default">New</a>
+                    <a href='{{ route('modal.user') }}' data-toggle="modal"  data-target="#createModal" data-dependency="user" data-select='assigned_user' class="btn btn-sm btn-default">New</a>
                   @endcan
               </div>
             </div>
@@ -158,142 +158,6 @@
 @stop
 
 @section('moar_scripts')
-
-    {{-- Some room for the modals --}}
-    <div class="modal fade" id="createModal">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                    <h4 class="modal-title">Modal title</h4>
-                </div>
-                <div class="modal-body">
-
-                    <div class="dynamic-form-row">
-                        <div class="col-md-4 col-xs-12"><label for="modal-first_name">{{ trans('general.first_name') }}:</label></div>
-                        <div class="col-md-8 col-xs-12 required"><input type='text' id='modal-first_name' class="form-control"></div>
-                    </div>
-
-                    <div class="dynamic-form-row">
-                        <div class="col-md-4 col-xs-12"><label for="modal-last_name">{{ trans('general.last_name') }}:</label></div>
-                        <div class="col-md-8 col-xs-12"><input type='text' id='modal-last_name' class="form-control"></div>
-                    </div>
-
-                    <div class="dynamic-form-row">
-                        <div class="col-md-4 col-xs-12"><label for="modal-username">{{ trans('admin/users/table.username') }}:</label></div>
-                        <div class="col-md-8 col-xs-12 required"><input type='text' id='modal-username' class="form-control"></div>
-                    </div>
-
-                    <div class="dynamic-form-row">
-                        <div class="col-md-4 col-xs-12"><label for="modal-email">{{ trans('admin/users/table.email') }}:</label></div>
-                        <div class="col-md-8 col-xs-12"><input type='email' id='modal-email' class="form-control"></div>
-                    </div>
-
-                    <div class="dynamic-form-row">
-                        <div class="col-md-4 col-xs-12"><label for="modal-password">{{ trans('admin/users/table.password') }}:</label></div>
-                        <div class="col-md-8 col-xs-12 required"><input type='password' id='modal-password' class="form-control"></div>
-                    </div>
-
-
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-default" data-dismiss="modal">{{ trans('button.cancel') }}</button>
-                    <button type="button" class="btn btn-primary" id="modal-save">{{ trans('general.save') }}</button>
-                </div>
-            </div><!-- /.modal-content -->
-        </div><!-- /.modal-dialog -->
-    </div><!-- /.modal -->
-
-
-
-
-    <!-- Ajax call to retrieve user assets -->
-
-<script>
-
-
-    $(function () {
-        var model,select;
-
-
-        $('#createModal').on("show.bs.modal",function (event) {
-            console.warn('modal ran');
-            var link = $(event.relatedTarget);
-            model=link.data("dependency");
-            select=link.data("select");
-
-            var modal = $(this);
-            modal.find('.modal-title').text('Add a new ' + model);
-
-            $('.dynamic-form-row').hide();
-            function show_er(selector) {
-                $(selector).parent().parent().show();
-            }
-            show_er('#modal-name');
-            switch(model) {
-
-                case 'user':
-                    $('.dynamic-form-row').hide(); //we don't want a generic "name"
-                    show_er("#modal-first_name");
-                    show_er("#modal-last_name");
-                    show_er("#modal-username");
-                    show_er("#modal-email");
-                    show_er("#modal-password");
-                    show_er("#modal-password_confirm");
-                    break;
-
-
-                    //do nothing, they just need 'name'
-            }
-
-            //console.warn("The Model is: "+model+" and the select is: "+select);
-        });
-
-
-
-
-
-        $('#modal-save').on('click',function () {
-            var data={};
-            //console.warn("We are about to SAVE!!! for model: "+model+" and select ID: "+select);
-            $('.modal-body input:visible').each(function (index,elem) {
-                //console.warn("["+index+"]: "+elem.id+" = "+$(elem).val());
-                var bits=elem.id.split("-");
-                if(bits[0]==="modal") {
-                    data[bits[1]]=$(elem).val();
-                }
-            });
-            $('.modal-body select:visible').each(function (index,elem) {
-                var bits=elem.id.split("-");
-                data[bits[1]]=$(elem).val();
-            });
-
-            data._token =  '{{ csrf_token() }}',
-                 //   console.dir(data);
-
-                $.post("{{url('/') }}/api/"+model+"s",data,function (result) {
-                    var id=result.id;
-                    var name=result.name || (result.first_name+" "+result.last_name);
-                    $('.modal-body input:visible').val("");
-                    $('#createModal').modal('hide');
-
-                    //console.warn("The select ID thing we're going for is: "+select);
-                    var selector=document.getElementById(select);
-                    selector.options[selector.length] = new Option(name,id);
-                    selector.selectedIndex=selector.length-1;
-                    $(selector).trigger("change");
-
-
-                }).fail(function (result) {
-                    //console.dir(result.responseJSON);
-                    msg=result.responseJSON.error.message || result.responseJSON.error;
-                    window.alert("Unable to add new "+model+" - error: "+msg);
-                });
-
-        });
-    });
-</script>
-
 <script>
 
 $(function() {
